XML-RPC: return HTTP response code when we have data to do so.
When calling secupress_die, one has the option to provide a response code in the array of arguments passed to the function.
That response code can be useful as it offers us a way to having an HTTP response code that matches the response we want to send.

It's especially important when using SecuPress' XML-RPC options in its Sensitive Data module.
That option allows us to block XML-RPC, which works well with the wp_die function.
However, the response code is only used and returned by WordPress when $wp_xmlrpc_server has been set:
https://github.com/WordPress/WordPress/blob/0050998801cabab21245452be315236d7d84da6b/wp-includes/functions.php#L3716
This is not the case here in secupress_die.
As a result, we end up blocking XML-RPC while still returning a 200 response code. This commit changes that and will return a 403 response code for XML-RPC requests made to a site using SecuPress.